requirements.txt
asgiref==3.9.2
cffi==2.0.0
cryptography==46.0.1
Django==5.2.6
django-cors-headers==4.3.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pycparser==2.23
PyJWT==2.10.1
python-decouple==3.8
sqlparse==0.5.3
